refactor: extract DB and redis secret fetchers
